+++ examples/foundational/01-say-one-thing.py
@@ -27,7 +27,8 @@ logger.remove(0)
 logger.add(sys.stderr, level="DEBUG")
 
 
-async def main(room_url):
+async def main():
+    (room_url, _) = await configure()
     async with aiohttp.ClientSession() as session:
         transport = DailyTransport(
             room_url, None, "Say One Thing", DailyParams(audio_out_enabled=True))
@@ -52,5 +53,4 @@ async def main(room_url):
         await runner.run(task)
 
 if __name__ == "__main__":
-    (url, token) = configure()
-    asyncio.run(main(url))
+    asyncio.run(main())