+++ examples/foundational/05-sync-speech-and-image.py
@@ -73,7 +73,8 @@ class MonthPrepender(FrameProcessor):
             await self.push_frame(frame, direction)
 
 
-async def main(room_url):
+async def main():
+    (room_url, _) = await configure()
     async with aiohttp.ClientSession() as session:
         transport = DailyTransport(
             room_url,
@@ -162,5 +163,4 @@ async def main(room_url):
 
 
 if __name__ == "__main__":
-    (url, token) = configure()
-    asyncio.run(main(url))
+    asyncio.run(main())