+++ examples/foundational/17-detect-user-idle.py
@@ -33,7 +33,9 @@ logger.remove(0)
 logger.add(sys.stderr, level="DEBUG")
 
 
-async def main(room_url: str, token):
+async def main():
+    (room_url, token) = await configure()
+
     async with aiohttp.ClientSession() as session:
         transport = DailyTransport(
             room_url,
@@ -104,5 +106,4 @@ async def main(room_url: str, token):
 
 
 if __name__ == "__main__":
-    (url, token) = configure()
-    asyncio.run(main(url, token))
+    asyncio.run(main())