+++ CHANGELOG.md
@@ -35,6 +35,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a `BaseInputTransport` issue that was causing start/stop interruptions
   incoming frames to not cancel tasks and be processed properly.
 
+### Other
+
+- Remove `requests` library usage.
+
+- Cleanup examples and use `DailyRESTHelper`.
+
 ## [0.0.39] - 2024-07-23
 
 ### Fixed